requirements.txt
pandas
numpy
scikit-learn
nltk
rich
typer
google-genai
python-dotenv
plotly
nbformat
streamlit
